product_service/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Migrate
